Fixed the problems with graphs plot when the programm was defined invalid steady mode points.
Now, the programm works correct when it was defined more than one invalid
points for any one temperature parameter.
Also if the programm was defined invalid points in every steady mode (even
if this points contains in different temperature parameters), all steady
mode points plots as invalid and prints INFO message in the console about it.

Signed-off-by: Oleg <olgerd27@gmail.com>